Further coding efficiencies and elegance.
